Organize scenarios into separate files
