+++ scenarios/scenario3.test.js
@@ -1,0 +1,90 @@
+/* Scenario 3: Tic-Tac-Toe Game with Leaderboard
+   Roles: player, moderator
+   Resources: games, leaderboard entries
+
+   Access Controls:
+   - Players may create a new game and invite one opponent.
+   - During an active game only the two participants may make moves.
+   - Completed games are read-only for all players.
+   - Leaderboard entries are generated automatically from game results.
+   - Players may read the leaderboard but cannot modify entries.
+   - Moderators may update or reset leaderboard entries and review detailed game history for dispute resolution.
+*/
+
+const assert = require('node:assert');
+const { test } = require('node:test');
+const { authorize } = require('../ruleEngine');
+
+const rules = [
+  {
+    when: { resource: 'game', action: 'create' },
+    rule: {
+      AND: [
+        { 'user.role': 'player' },
+        { 'user.id': { in: { reference: 'item.participants' } } }
+      ]
+    }
+  },
+  {
+    when: { resource: 'game', action: 'move' },
+    rule: {
+      AND: [
+        { 'user.id': { in: { reference: 'item.participants' } } },
+        { 'item.status': { not: 'complete' } }
+      ]
+    }
+  },
+  {
+    when: { resource: 'game', action: 'read' },
+    rule: {
+      OR: [
+        { 'item.status': 'complete' },
+        {
+          AND: [
+            { 'user.id': { in: { reference: 'item.participants' } } },
+            { 'item.status': { not: 'complete' } }
+          ]
+        }
+      ]
+    }
+  },
+  {
+    when: { resource: 'leaderboard', action: 'read' },
+    rule: { 'user.role': { in: ['player', 'moderator'] } }
+  },
+  {
+    when: { resource: 'leaderboard', action: 'update' },
+    rule: { 'user.role': 'moderator' }
+  }
+];
+
+module.exports = { rules };
+
+// Tests
+
+test('scenario3: participant can move in active game', () => {
+  const context = {
+    resource: 'game',
+    action: 'move',
+    user: { id: 'p1', role: 'player' },
+    item: { participants: ['p1', 'p2'], status: 'active' }
+  };
+  assert.strictEqual(authorize(rules, context), true);
+});
+
+test('scenario3: non participant cannot move', () => {
+  const context = {
+    resource: 'game',
+    action: 'move',
+    user: { id: 'x', role: 'player' },
+    item: { participants: ['p1', 'p2'], status: 'active' }
+  };
+  assert.strictEqual(authorize(rules, context), false);
+});
+
+test('scenario3: only moderator updates leaderboard', () => {
+  const context = { resource: 'leaderboard', action: 'update', user: { role: 'player' } };
+  assert.strictEqual(authorize(rules, context), false);
+  const modCtx = { resource: 'leaderboard', action: 'update', user: { role: 'moderator' } };
+  assert.strictEqual(authorize(rules, modCtx), true);
+});
